dodawanie i usuwanie raportow

+++ src/app/api/audits/[id]/route.js
@@ -1,15 +1,26 @@
-import { NextResponse } from 'next/server'
 import { prisma } from '@/lib/prisma'
+import { getServerSession } from 'next-auth'
+import { authOptions } from '@/lib/authOptions'
+import { NextResponse } from 'next/server'
 
-export async function DELETE(_, { params }) {
+export async function DELETE(req, { params }) {
+	const session = await getServerSession(authOptions)
+
+	if (!session || session.user.role !== 'ADMIN') {
+		return new NextResponse('Unauthorized', { status: 401 })
+	}
+
 	const { id } = params
 
 	try {
-		await prisma.file.deleteMany({ where: { auditId: id } }) // usuń powiązane pliki
+		await prisma.error.deleteMany({ where: { auditId: id } })
+		await prisma.file.deleteMany({ where: { auditId: id } })
+		await prisma.report.deleteMany({ where: { auditId: id } })
 		await prisma.audit.delete({ where: { id } })
 
-		return NextResponse.json({ success: true })
+		return new NextResponse(null, { status: 204 })
 	} catch (error) {
-		return new NextResponse('Błąd przy usuwaniu audytu', { status: 500 })
+		console.error('Błąd usuwania audytu:', error)
+		return new NextResponse('Błąd serwera', { status: 500 })
 	}
 }
